Fix and improve tests that were affected by files that were removed
