+++ tests/utils/analysis_data.py
@@ -1,94 +1,140 @@
-from tests.test_helpers import create_file_df
-
-JSON_LIST = [
-    "tests/unit/data/fga-eps-mds-2021-2-MeasureSoftGram-Service-04-12-2022-17-32-35-v1.1.0.json"
-]
-
-SUBS = {
-    "testing_status": {
-        "measures": ["test_builds", "test_coverage", "passed_tests"],
-        "weights": {
-            "test_builds": 0.3333,
-            "test_coverage": 0.3333,
-            "passed_tests": 0.3333,
-        },
-    }
-}
-
-MEASURES = {
-    "test_builds": 0.0010166666666666666,
-    "test_coverage": 0.5647222222222222,
-    "passed_tests": 0.7142857142857143,
-}
-
-CHARACTERISTICS = {
-    "reliability": {
-        "weight": 1.0,
-        "subcharacteristics": ["testing_status"],
-        "weights": {"testing_status": 1.0},
-    },
-}
-
-FILES_DF = create_file_df(
-    JSON_LIST,
-    "py",
-)
-
-CALCULATE_MEASURES_DATA = [
-    (
-        "test_builds",
-        {
+EXTRACTED_MEASURES_DATA = {
+    'measures':
+    [
+        {'key': 'passed_tests', 'parameters': {
+            'tests': [
+                3.0, 1.0, 6.0, 3.0, 17.0, 2.0, 2.0, 7.0, 1.0, 1.0, 1.0, 3.0, 2.0,
+                3.0, 23.0, 2.0, 3.0, 4.0, 7.0, 2.0
+            ], 'test_failures': 0.0, 'test_errors': 0.0}},
+        {'key': 'test_builds', 'parameters': {
             'test_execution_time': [
-                6.0, 2.0, 18.0, 6.0, 17.0, 969.0, 4.0, 9.0, 1.0, 961.0, 476.0, 5.0, 954.0,
-                3.0, 24.0, 23.0, 3.0, 7.0, 47.0, 5.0
+                6.0, 2.0, 18.0, 6.0, 17.0, 969.0, 4.0, 9.0, 1.0, 961.0, 476.0, 5.0,
+                954.0, 3.0, 24.0, 23.0, 3.0, 7.0, 47.0, 5.0
             ],
             'tests': [
-                3.0, 1.0, 6.0, 3.0, 17.0, 2.0, 2.0, 7.0, 1.0, 1.0, 1.0, 3.0, 2.0, 3.0,
-                23.0, 2.0, 3.0, 4.0, 7.0, 2.0
+                3.0, 1.0, 6.0, 3.0, 17.0, 2.0, 2.0, 7.0, 1.0, 1.0, 1.0, 3.0, 2.0,
+                3.0, 23.0, 2.0, 3.0, 4.0, 7.0, 2.0
             ]
-        },
-        0.9999959333997583,
-    ),
-    (
-        "test_coverage",
-        {'coverage': [
+        }},
+        {'key': 'test_coverage', 'parameters': {'coverage': [
             0.0, 0.0, 0.0, 0.0, 0.0, 44.4, 23.5, 100.0, 100.0, 50.0, 77.8, 100.0,
             100.0, 91.4, 100.0, 43.8, 0.0, 55.6, 26.7, 100.0, 100.0, 100.0, 0.0,
             100.0, 64.7, 86.6, 0.0, 61.9, 91.8, 94.4, 82.5, 13.3
-        ]},
-        0.4515625,
+        ]}},
+        {'key': 'non_complex_file_density', 'parameters': {
+            'functions': [
+                0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
+                0.0, 0.0, 0.0, 0.0, 3.0, 3.0, 2.0, 1.0, 1.0, 0.0, 8.0, 10.0, 1.0, 1.0,
+                1.0, 1.0, 1.0, 3.0, 1.0, 3.0, 1.0, 1.0, 1.0, 1.0, 3.0, 7.0, 7.0, 2.0,
+                4.0, 4.0, 14.0, 6.0, 3.0
+            ],
+            'complexity': [
+                0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
+                0.0, 0.0, 0.0, 0.0, 4.0, 4.0, 6.0, 1.0, 2.0, 0.0, 14.0, 23.0, 1.0, 6.0,
+                2.0, 2.0, 1.0, 11.0, 5.0, 15.0, 1.0, 4.0, 5.0, 1.0, 13.0, 13.0, 19.0,
+                3.0, 7.0, 11.0, 51.0, 19.0, 10.0
+            ]}},
+        {'key': 'commented_file_density', 'parameters': {'comment_lines_density': [
+            0.0, 0.0, 0.0, 5.7, 2.1, 4.8, 0.0, 0.0, 38.1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
+            0.0, 8.5, 0.0, 5.9, 0.0, 0.0, 0.0, 6.4, 0.0, 1.4, 14.8, 16.7, 35.3, 0.0,
+            0.0, 0.0, 0.0
+        ]}},
+        {'key': 'duplication_absense', 'parameters': {'duplicated_lines_density': [
+            0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
+            0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
+            0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
+            0.0, 0.0
+        ]}}]
+}
+
+CALCULATE_MEASURES_RESPONSE_DATA = {
+    'measures':
+    [
+        {'key': 'passed_tests', 'value': 1.0},
+        {'key': 'test_builds', 'value': 0.9999959333997583},
+        {'key': 'test_coverage', 'value': 0.4515625},
+        {'key': 'non_complex_file_density', 'value': 0.5361702127659572},
+        {'key': 'commented_file_density', 'value': 0.04453125},
+        {'key': 'duplication_absense', 'value': 1.0}
+    ]
+}
+
+
+CALCULATE_MEASURES_ERROR_INFOS = [
+    (
+        {'measures': None},
+        "Failed to validate request"
     ),
     (
-        "passed_tests",
-        {'test_failures': 0.0, 'test_errors': 0.0, 'tests': [
-            3.0, 1.0, 6.0, 3.0, 17.0, 2.0, 2.0, 7.0, 1.0, 1.0, 1.0, 3.0, 2.0, 3.0, 23.0, 2.0, 3.0, 4.0, 7.0, 2.0]},
-        1.0
+        {'measures': [{'key': 'inexistent', 'parameters': {'inexistent_v2': [1]}}]},
+        "Measure inexistent is not supported"
+    ),
+    (
+        {'measures': [{'key': 'duplication_absense', 'parameters': {'duplicated_lines_density': [None]}}]},
+        "Metric parameters duplication_absense are not valid"
     ),
 ]
 
-RESOLVE_LEVEL_DATA = [
+EXTRACTED_SUBCHARACTERISTICS_DATA = {
+    'subcharacteristics': [
+        {'key': 'testing_status', 'measures': [
+            {'key': 'passed_tests', 'value': 1.0, 'weight': 33},
+            {'key': 'test_builds', 'value': 0.9999959333997583, 'weight': 33},
+            {'key': 'test_coverage', 'value': 0.4515625, 'weight': 34}]},
+        {'key': 'modifiability', 'measures': [
+            {'key': 'non_complex_file_density', 'value': 0.5361702127659572, 'weight': 33},
+            {'key': 'commented_file_density', 'value': 0.04453125, 'weight': 33},
+            {'key': 'duplication_absense', 'value': 1.0, 'weight': 34}]}
+    ]
+}
+CALCULATE_SUBCHARACTERISTICS_RESPONSE_DATA = {
+    'subcharacteristics': [
+        {'key': 'testing_status', 'value': 0.8508631402592323},
+        {'key': 'modifiability', 'value': 0.6642882099299446}
+    ]
+}
+CALCULATE_SUBCHARACTERISTICS_ERROR_INFOS = [
     (
-        SUBS,
-        MEASURES,
-        "measures",
-        0.525711,
-    ),
+        {'subcharacteristics': None},
+        "Failed to validate request"
+    )
 ]
 
-MAKE_ANALYSIS_DATA = [
+EXTRACTED_CHARACTERISTICS_DATA = {
+    'characteristics': [
+        {'key': 'reliability', 'subcharacteristics': [
+            {'key': 'testing_status', 'value': 0.8508631402592323, 'weight': 100}]},
+        {'key': 'maintainability', 'subcharacteristics': [
+            {'key': 'modifiability', 'value': 0.6642882099299446, 'weight': 100}]}
+    ]
+}
+CALCULATE_CHARACTERISTICS_RESPONSE_DATA = {
+    'characteristics': [
+        {'key': 'reliability', 'value': 0.8508631402592323},
+        {'key': 'maintainability', 'value': 0.6642882099299446}
+    ]
+}
+CALCULATE_CHARACTERISTICS_ERROR_INFOS = [
     (
-        MEASURES,
-        SUBS,
-        CHARACTERISTICS,
-        [
-            0.525711,
-            0.525711,
-            0.525711,
-        ],
-        [
-            "testing_status",
-            "reliability",
-            "sqc",
-        ]
-    ),
+        {'characteristics': None},
+        "Failed to validate request"
+    )
+]
+
+EXTRACTED_SQC_DATA = {
+    'sqc': {'key': 'sqc', 'characteristics': [
+        {'key': 'reliability', 'value': 0.8508631402592323, 'weight': 50},
+        {'key': 'maintainability', 'value': 0.6642882099299446, 'weight': 50}
+    ]}
+}
+CALCULATE_SQC_RESPONSE_DATA = {
+    'sqc': [
+        {'key': 'sqc', 'value': 0.76329774967038}
+    ]
+}
+CALCULATE_SQC_ERROR_INFOS = [
+    (
+        {'sqc': None},
+        "Failed to validate request"
+    )
 ]
