+++ tox.ini
@@ -21,6 +21,6 @@ setenv =
 
 commands =
     coverage erase
-    coverage run -m pytest {posargs} tests --junitxml=./junit.xml
+    coverage run -m pytest tests {posargs} --junitxml=./junit.xml
     coverage xml -o coverage.xml
-    coverage report
+    coverage report --show-missing
